requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.22.0
pandas==1.4.2
psycopg==3.0.14
psycopg2==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
whitenoise==6.1.0
